Respostas/requirements.txt
numpy
aiohttp
matplotlib
cython
Pillow
